Share code between extract and diff
