+++ src/vorta/views/diff_result.py
@@ -2,31 +2,25 @@ import os
 import re
 
 from PyQt5 import uic
-from PyQt5.QtCore import QAbstractItemModel, QModelIndex, Qt, QVariant
+from PyQt5.QtCore import Qt, QVariant
 from PyQt5.QtGui import QColor
 from PyQt5.QtWidgets import QHeaderView
 
-from vorta.utils import (get_asset, get_dict_from_list, nested_dict,
-                         pretty_bytes)
+from vorta.utils import (get_asset, get_dict_from_list, nested_dict)
+
+from vorta.views.tree_view import TreeModel
 
 uifile = get_asset('UI/diffresult.ui')
 DiffResultUI, DiffResultBase = uic.loadUiType(uifile)
-
-files_with_attributes = None
-nested_file_list = None
-selected_files_folders = None
 
 
 class DiffResult(DiffResultBase, DiffResultUI):
     def __init__(self, fs_data, archive_newer, archive_older):
         super().__init__()
         self.setupUi(self)
-        global files_with_attributes, nested_file_list, selected_files_folders
 
-        # Clear global file lists
         files_with_attributes = []
         nested_file_list = nested_dict()
-        selected_files_folders = set()
 
         def parse_line(line):
             if line:
@@ -77,7 +71,7 @@ class DiffResult(DiffResultBase, DiffResultUI):
         for l in fs_data.split('\n'):
             files_with_attributes.append(parse_line(l))
 
-        model = TreeModel()
+        model = DiffTree(files_with_attributes, nested_file_list)
 
         view = self.treeView
         view.setAlternatingRowColors(True)
@@ -92,7 +86,6 @@ class DiffResult(DiffResultBase, DiffResultUI):
         self.archiveNameLabel_1.setText(f'{archive_newer.name}')
         self.archiveNameLabel_2.setText(f'{archive_older.name}')
         self.okButton.clicked.connect(self.accept)
-        self.selected = selected_files_folders
 
 
 def calc_size(significand, unit):
@@ -108,110 +101,16 @@ def calc_size(significand, unit):
         return int(float(significand) * 10**12)
 
 
-class FolderItem:
-    def __init__(self, path, name, modified, parent=None):
-        self.parentItem = parent
-        self.path = path
-        self.itemData = [name, modified]
-        self.childItems = []
-
-        # Pre-filter children
-        self._filtered_children = []
-        search_path = os.path.join(self.path, name)
-        if parent is None:  # Find path for root folder
-            for root_folder in nested_file_list.keys():
-                self._filtered_children.append((0, '', root_folder, '', ))
-        else:
-
-            # This adds direct children.
-            self._filtered_children = [f for f in files_with_attributes if search_path == f[3]]
-
-            # Add nested folders.
-            for immediate_child in get_dict_from_list(nested_file_list, search_path.split('/')).keys():
-                if not [True for child in self._filtered_children if child[2] == immediate_child]:
-                    self._filtered_children.append((0, '', immediate_child, search_path))
-
-        self.is_loaded = False
-
-    def load_children(self):
-        for child_item in self._filtered_children:
-            if child_item[0] > 0:  # This is a file
-                self.childItems.append(FileItem(
-                    name=child_item[2],
-                    modified=child_item[1],
-                    size=child_item[0],
-                    parent=self))
-            else:  # Folder
-                self.childItems.append(
-                    FolderItem(
-                        path=child_item[3],
-                        name=child_item[2],
-                        modified=child_item[1],
-                        parent=self))
-        self.is_loaded = True
-
-    def child(self, row):
-        return self.childItems[row]
-
-    def childCount(self):
-        return len(self._filtered_children)
-
-    def columnCount(self):
-        return 3
-
-    def data(self, column):
-        if column <= 1:
-            return self.itemData[column]
-        else:
-            return None
-
-    def parent(self):
-        return self.parentItem
-
-    def row(self):
-        if self.parentItem:
-            return self.parentItem.childItems.index(self)
-
-        return 0
-
-
-class FileItem(FolderItem):
-    def __init__(self, name, modified, size, parent=None):
-        self.parentItem = parent
-        self.itemData = [name, modified, size]
-
-    def childCount(self):
-        return 0
-
-    def columnCount(self):
-        return 3
-
-    def data(self, column):
-        if column == 1:
-            return self.itemData[column]
-        elif column == 2:
-            return pretty_bytes(self.itemData[column])
-        elif column == 0:
-            return self.itemData[column]
-
-    def parent(self):
-        return self.parentItem
-
-    def row(self):
-        return self.parentItem.childItems.index(self)
-
-
-class TreeModel(QAbstractItemModel):
-    column_names = ['Name', 'Modified', 'Size']
-
-    def __init__(self, parent=None):
-        super(TreeModel, self).__init__(parent)
-
-        self.rootItem = FolderItem(path='', name='', modified=None)
-        self.rootItem.load_children()
-
-    def columnCount(self, parent):
-        return 3
+class DiffTree(TreeModel):
+    def __init__(
+        self,
+        files_with_attributes,
+        nested_file_list,
+        parent=None,
+    ):
+        super().__init__(
+            files_with_attributes, nested_file_list, parent=parent
+        )
 
     def data(self, index, role):
         if not index.isValid():
@@ -232,61 +131,7 @@ class TreeModel(QAbstractItemModel):
         else:
             return None
 
-    def canFetchMore(self, index):
-        if not index.isValid():
-            return False
-        item = index.internalPointer()
-        return not item.is_loaded
-
-    def fetchMore(self, index):
-        item = index.internalPointer()
-        item.load_children()
-
     def flags(self, index):
         if not index.isValid():
             return Qt.NoItemFlags
         return Qt.ItemIsEnabled
-
-    def headerData(self, section, orientation, role):
-        if orientation == Qt.Horizontal and role == Qt.DisplayRole:
-            return self.column_names[section]
-
-        return None
-
-    def index(self, row, column, parent):
-        if not self.hasIndex(row, column, parent):
-            return QModelIndex()
-
-        if not parent.isValid():
-            parentItem = self.rootItem
-        else:
-            parentItem = parent.internalPointer()
-
-        childItem = parentItem.child(row)
-        if childItem:
-            return self.createIndex(row, column, childItem)
-        else:
-            return QModelIndex()
-
-    def parent(self, index):
-        if not index.isValid():
-            return QModelIndex()
-
-        childItem = index.internalPointer()
-        parentItem = childItem.parent()
-
-        if parentItem == self.rootItem:
-            return QModelIndex()
-
-        return self.createIndex(parentItem.row(), 0, parentItem)
-
-    def rowCount(self, parent):
-        if parent.column() > 0:
-            return 0
-
-        if not parent.isValid():
-            parentItem = self.rootItem
-        else:
-            parentItem = parent.internalPointer()
-
-        return parentItem.childCount()
